Refactor Edit modal
