+++ web/app/components/document/sidebar/related-resources/list-item/edit.ts
@@ -1,9 +1,8 @@
-import { assert } from "@ember/debug";
 import { action } from "@ember/object";
+import { guidFor } from "@ember/object/internals";
 import Component from "@glimmer/component";
 import { tracked } from "@glimmer/tracking";
 import { RelatedExternalLink } from "hermes/components/document/sidebar/related-resources";
-import isValidURL from "hermes/utils/is-valid-u-r-l";
 
 interface DocumentSidebarRelatedResourcesListItemEditComponentSignature {
   Args: {
@@ -18,101 +17,14 @@ interface DocumentSidebarRelatedResourcesListItemEditComponentSignature {
 }
 
 export default class DocumentSidebarRelatedResourcesListItemEditComponent extends Component<DocumentSidebarRelatedResourcesListItemEditComponentSignature> {
-  /**
-   * A locally tracked URL property. Starts as the passed-in value;
-   * updated when the URL input-value changes.
-   */
-  @tracked protected url = this.args.resource ? this.args.resource.url : "";
+  protected id = guidFor(this);
+  protected bodySelector = `#${this.id}-body`;
+  protected submitSelector = `#${this.id}-submit`;
 
-  /**
-   * The title of the resource. If the name is the same as the URL,
-   * we treat it like it's an empty value so the placeholder text shows.
-   */
-  @tracked protected title = !this.args.resource
-    ? ""
-    : this.args.resource.name === this.args.resource.url
-    ? ""
-    : this.args.resource.name;
+  @tracked protected formIsRendered = false;
 
-  /**
-   * Whether the URL is valid, as determined by the `isValidURL` utility.
-   * Used to dictate whether an error message is shown.
-   */
-  @tracked protected urlIsValid = true;
-
-  /**
-   * Whether the error warning is shown.
-   * True if the title is empty on submit.
-   */
-  @tracked protected titleErrorIsShown = false;
-
-  /**
-   * A local reference to the form element.
-   * Registered when inserted.
-   */
-  @tracked private _form: HTMLFormElement | null = null;
-
-  /**
-   * An asserted-true reference to the form element.
-   */
-  protected get form(): HTMLFormElement {
-    assert("this._form must exist", this._form);
-    return this._form;
-  }
-
-  /**
-   * The action to register the form element locally.
-   * Called when the form is rendered.
-   */
-  @action protected registerForm(form: HTMLFormElement): void {
-    this._form = form;
-  }
-
-  /**
-   * The action that updates the local form properties and eagerly validates the URL.
-   * Runs on the "input" events of the URL and title inputs.
-   */
-  @action protected updateFormValues(): void {
-    const formObject = Object.fromEntries(new FormData(this.form).entries());
-
-    const title = formObject["title"];
-    const url = formObject["url"];
-
-    assert("title must be a string", typeof title === "string");
-    assert("url must be a string", typeof url === "string");
-
-    this.title = title;
-    this.url = url;
-
-    this.validateURL();
-  }
-
-  /**
-   * The action that validates the locally tracked URL.
-   * Updates the `urlIsValid` property, which is used to show
-   * an error message for invalid URLs.
-   */
-  @action private validateURL() {
-    this.urlIsValid = isValidURL(this.url);
-  }
-
-  @action maybeSave() {
-    this.validateURL();
-
-    if (!this.title) {
-      this.titleErrorIsShown = true;
-      return;
-    }
-
-    if (this.urlIsValid) {
-      if (this.args.onSave) {
-        this.args.onSave({
-          sortOrder: this.args.resource.sortOrder,
-          name: this.title,
-          url: this.url,
-        });
-      }
-    }
+  @action protected renderForm() {
+    this.formIsRendered = true;
   }
 }
